requirements.txt
backcall==0.1.0
decorator==4.3.0
flake8==3.5.0
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.12.1
kafka-python==1.4.3
mccabe==0.6.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.4
ptyprocess==0.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pymongo==3.7.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
